+++ src/xyz/oihana/schema/constants/traits/business/documents/BusinessDocumentTrait.php
@@ -31,5 +31,6 @@ trait BusinessDocumentTrait
     public const string STATUS          = 'status' ;
     public const string TAXES           = 'taxes' ;
     public const string TOTALS          = 'totals' ;
+    public const string VOLUME          = 'volume' ;
     public const string WEIGHT          = 'weight' ;
 }